Bug fixes for debug mode.
I've not been seeing compile errors for debug mode because I'm now using `llvm-config --cxxflags` in CMakeLists.txt with a release-mode LLVM, so my CXXFLAGS contains "-D DEBUG [...] -O3 -DNDEBUG"!

After fixing my CXXFLAGS, I can see some compile errors that need fixing:
 * ASSERTION_FN_NAME -> INLINE_ASSERTION (changed awhile ago)
 * some signed/unsigned comparisons
 * comparing an OwningPtr to NULL

+++ tesla/analyser/Tool.cpp
@@ -67,7 +67,7 @@ int main(int argc, const char **argv) {
   // The CompilationDatabase will modify these, so we shouldn't pass in
   // args.data() directly.
   argc = (int) args.size();
-  assert(argc == args.size());    // check for overflow
+  assert(((size_t) argc) == args.size());    // check for overflow
 
   argv = args.data();
